gen-ai/inference/nvidia-triton-server-gpu/yolo/triton-client/requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.0
aiohttp==3.10.5
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.4.0
attrs==24.2.0
blinker==1.8.2
Brotli==1.1.0
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
ConfigArgParse==1.7
contourpy==1.3.0
cycler==0.12.1
fastapi==0.112.2
ffmpy==0.4.0
filelock==3.15.4
Flask==3.0.3
Flask-Cors==5.0.0
Flask-Login==0.6.3
fonttools==4.53.1
frozenlist==1.4.1
fsspec==2024.6.1
gevent==24.2.1
geventhttpclient==2.3.1
gradio==4.42.0
gradio_client==1.3.0
greenlet==3.0.3
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface-hub==0.24.6
idna==3.8
importlib_resources==6.4.4
itsdangerous==2.2.0
Jinja2==3.1.4
kiwisolver==1.4.5
locust==2.31.5
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
networkx==3.3
numpy==1.26.4
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.2
pillow==10.4.0
protobuf==3.20.3
psutil==6.0.0
py-cpuinfo==9.0.0
pydantic==2.8.2
pydantic_core==2.20.1
pydub==0.25.1
Pygments==2.18.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-multipart==0.0.9
python-rapidjson==1.20
pytz==2024.1
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.8.0
ruff==0.6.3
scipy==1.14.1
seaborn==0.13.2
semantic-version==2.10.0
setuptools==74.1.0
shellingham==1.5.4
six==1.16.0
sniffio==1.3.1
starlette==0.38.4
sympy==1.13.2
tomlkit==0.12.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.5
tritonclient==2.36.0
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.1
ultralytics==8.2.87
ultralytics-thop==2.0.6
urllib3==2.2.2
uvicorn==0.30.6
websockets==12.0
Werkzeug==3.0.4
yarl==1.9.7
zope.event==5.0
zope.interface==7.0.3
